worker/requirements.txt
google-cloud-pubsub==2.23.0
PyYAML==6.0.1
requests==2.32.2